Bot/requirements.txt
python-telegram-bot~=13.5
owlready2~=0.39